:white_check_mark: feat: 初步完成tile instance的开发，缺少tile api的调试(namespace,datastore,feature layer的crud均调试通过)

+++ component/metadata/src/main/java/org/geoatlas/metadata/persistence/managent/DataStoreInfoManagement.java
@@ -2,9 +2,17 @@ package org.geoatlas.metadata.persistence.managent;
 
 import org.geoatlas.metadata.GeoAtlasMetadataContext;
 import org.geoatlas.metadata.model.DataStoreInfo;
+import org.geoatlas.metadata.model.NamespaceInfo;
 import org.geoatlas.metadata.persistence.repository.DataStoreInfoRepository;
 import org.jasypt.encryption.pbe.PBEStringEncryptor;
+import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.CollectionUtils;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 /**
  * @author: <a href="mailto:thread.zhou@gmail.com">Fuyi</a>
@@ -16,29 +24,53 @@ public class DataStoreInfoManagement {
 
     private final PBEStringEncryptor encryptor;
     private final DataStoreInfoRepository repository;
+    private final NamespaceInfoManagement namespaceInfoManagement;
+    private final ModelMapper mapper;
 
-    public DataStoreInfoManagement(PBEStringEncryptor encryptor, DataStoreInfoRepository repository) {
+    public DataStoreInfoManagement(PBEStringEncryptor encryptor, DataStoreInfoRepository repository,
+                                   NamespaceInfoManagement namespaceInfoManagement, ModelMapper mapper) {
         this.encryptor = encryptor;
         this.repository = repository;
+        this.namespaceInfoManagement = namespaceInfoManagement;
+        this.mapper = mapper;
     }
 
+    @Transactional
     public void addDataStoreInfo(DataStoreInfo info) {
+        NamespaceInfo namespaceInfo = namespaceCheck(info);
+        // 先删除后新增, 保证一个Namespace下只有一个Datastore
+        List<DataStoreInfo> badData = repository.findDataStoreInfoByNamespaceId(namespaceInfo.getId());
+        if (!CollectionUtils.isEmpty(badData)) {
+            repository.deleteAllById(badData.stream().map(DataStoreInfo::getId).collect(Collectors.toList()));
+        }
         info.setPassword(encryptor.encrypt(info.getPassword()));
         DataStoreInfo saved = repository.save(info);
         saved.setPassword(encryptor.decrypt(saved.getPassword()));
-        GeoAtlasMetadataContext.addDataStore(saved);
+        GeoAtlasMetadataContext.addDataStore(namespaceInfo.getName(), saved);
     }
 
+    @Transactional
     public void updateDataStoreInfo(DataStoreInfo info) {
-        DataStoreInfo saved = repository.save(info);
-        saved.setPassword(encryptor.decrypt(saved.getPassword()));
-        GeoAtlasMetadataContext.removeDataStore(saved.getId());
-        GeoAtlasMetadataContext.addDataStore(saved);
+        Optional<DataStoreInfo> oldOptional = repository.findById(info.getId());
+        if (oldOptional.isPresent()){
+            NamespaceInfo namespaceInfo = namespaceCheck(info);
+            DataStoreInfo last = oldOptional.get();
+            mapper.map(info, last);
+            info.setPassword(encryptor.encrypt(info.getPassword()));
+            DataStoreInfo saved = repository.save(info);
+            saved.setPassword(encryptor.decrypt(saved.getPassword()));
+            GeoAtlasMetadataContext.removeDataStore(namespaceInfo.getName());
+            GeoAtlasMetadataContext.addDataStore(namespaceInfo.getName(), saved);
+        }
     }
 
     public void removeDataStoreInfo(Long id) {
-        repository.deleteById(id);
-        GeoAtlasMetadataContext.removeDataStore(id);
+        Optional<DataStoreInfo> target = repository.findById(id);
+        if(target.isPresent()) {
+            NamespaceInfo namespaceInfo = namespaceCheck(target.get());
+            repository.deleteById(id);
+            GeoAtlasMetadataContext.removeDataStore(namespaceInfo.getName());
+        }
     }
 
     /**
@@ -48,6 +80,14 @@ public class DataStoreInfoManagement {
      */
     public DataStoreInfo getDataStoreInfo(Long id) {
         return repository.findById(id).orElse(null);
+    }
+
+    private NamespaceInfo namespaceCheck(DataStoreInfo info){
+        NamespaceInfo namespaceInfo = namespaceInfoManagement.getNamespaceInfo(info.getNamespaceId());
+        if (namespaceInfo == null) {
+            throw new RuntimeException("namespace not found");
+        }
+        return namespaceInfo;
     }
 
 }
